requirements.txt
albumentations==2.0.8
hydra-core==1.3.2
omegaconf==2.3.0
opencv-python-headless==4.12.0.88
scikit-multilearn==0.2.0
numpy==2.2.6
pandas==2.3.3
pytorch-lightning==2.5.5
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
timm==1.0.20